Update tests
Removed broken Card component tests because of Drag and Drop. Add tests
for Cribbage app for moving cards around, and a basic snapshot test.
Update version of jest. Move asynchronous getting of deck to app.tsx
instead of Cribbage. This was done to allow Cribbage to be a bit more
standalone and anything could feed it the deck it wants it to use.